完善性修改
