+++ app/src/main/res/values/com1314080901139colors.xml
@@ -9,5 +9,6 @@
     <color name="bg_color_blue">#43D5FF</color>
     <color name="bg_color_blue_light">#84B5CA</color>
 
+    <color name="btn_color_init">#2FAB4D</color>
     <color name="text_color_r">#00CBBB</color>
 </resources>
